Group address and value into item
This groups the address and value into an item.
The idea is that most part of the merkle tree is independend of the
structure of the items in the leaves, so they should just operate on an
item. The only part where the structure matters is when calculating the
leave hashes. This should also make it simple to switch the format if it
changes.