Return '200 OK' for non entitled addresses
Previously we returned a 404 error when addresses where not contained
in the merkle tree.

This is indistinguishable from trying to load from a wrong server/URL
and may hide those errors from the user.

Instead we now return a normal response with originalTokenBalance set
to zero and an empty proof.