Use reorder-python-imports pre-commit hook
This sorts and groups imports.

One thing it does may be controversial. It splits 'from imports', e.g.

    from os.path import abspath, exists

becomes

    from os.path import abspath
    from os.path import exists

It does this because it tries to minimize merge conflicts.

Another alternative to sorting imports is isort, which the raiden
project is using.

isort didn't work out of the box, because it formats files not
compatible with black's formatting.

Visit https://github.com/asottile/reorder_python_imports#what-does-it-do
for more information.