Relax python version used by pre-commit
This relaxes the python version required and allows us to use the
pre-commit hooks on machines where only python 3.7 is available.

We still need at least python 3.6 for black, but there's no way to
declare that.

Also use the `default_language_version` in order to only having to
specify this once.