Remove withdrawFor function on MerkelDrop contract
Only the eligible person should be able to claim tokens from the
MerkleDrop contract for himself.
The tests were so far focused on the `withdrawFor` function, since it
simplified process. Since the eligible person needs to sign the
transaction himself, the used addresses for the test merkle-drop were in
need of some ETH to pay for the gas. Therefore the eth-tester test
accounts have replaced the original eligible address list.

Related to: #59