Fix declared dependencies
The web3 and contract-deploy-tools dependencies had been missing. The
eth_hash[pycryptodome] dependency wasn't used.

We now add gunicorn as an explicit dependency, because we recommend
that for running the backend server. That's one thing less, that can
go wrong (even if it's just a single command to install it).